ManagedRun cleanup
