+++ core/src/worker/workflow/run_cache.rs
@@ -1,20 +1,20 @@
 use crate::{
     telemetry::metrics::workflow_type,
     worker::workflow::{
-        managed_run::WorkflowManager, HeartbeatTimeoutMsg, HistoryUpdate, LocalActivityRequestSink,
-        ManagedRunHandle, NewIncomingWFT, RunUpdateResponseKind,
+        managed_run::ManagedRun, ActivationOrAuto, HeartbeatTimeoutMsg, HistoryUpdate,
+        LocalActivityRequestSink, PermittedWFT,
     },
     MetricsContext,
 };
 use lru::LruCache;
-use std::{num::NonZeroUsize, time::Instant};
+use std::{mem, num::NonZeroUsize};
 use tokio::sync::mpsc::UnboundedSender;
 
 pub(super) struct RunCache {
     max: usize,
     namespace: String,
     /// Run id -> Data
-    runs: LruCache<String, ManagedRunHandle>,
+    runs: LruCache<String, ManagedRun>,
     // TODO: eliminate & make return value
     local_activity_request_sink: LocalActivityRequestSink,
     // Used to feed heartbeat timeout events back into the workflow inputs.
@@ -52,71 +52,52 @@ impl RunCache {
         }
     }
 
-    pub fn instantiate_or_update(
-        &mut self,
-        run_id: &str,
-        workflow_id: &str,
-        wf_type: &str,
-        history_update: HistoryUpdate,
-        start_time: Instant,
-    ) -> (&mut ManagedRunHandle, RunUpdateResponseKind) {
+    pub fn instantiate_or_update(&mut self, mut pwft: PermittedWFT) -> Vec<ActivationOrAuto> {
         let cur_num_cached_runs = self.runs.len();
+        let run_id = &pwft.work.execution.run_id;
 
-        if self.runs.contains(run_id) {
-            // For some weird reason, maybe a NLL bug, there are double-mutable-borrow errors if I
-            // use get_mut above instead of in here (even though we always return from this branch).
-            // So, forced to do this.
-            let run_handle = self.runs.get_mut(run_id).unwrap();
-
-            run_handle.metrics.sticky_cache_hit();
-            let rur = run_handle.incoming_wft(NewIncomingWFT {
-                history_update: Some(history_update),
-                start_time,
-            });
+        if let Some(run_handle) = self.runs.get_mut(run_id) {
+            let rur = run_handle.incoming_wft(pwft);
             self.metrics.cache_size(cur_num_cached_runs as u64);
-            return (run_handle, rur);
+            return rur;
         }
 
         // Create a new workflow machines instance for this workflow, initialize it, and
         // track it.
         let metrics = self
             .metrics
-            .with_new_attrs([workflow_type(wf_type.to_string())]);
-        let wfm = WorkflowManager::new(
+            .with_new_attrs([workflow_type(pwft.work.workflow_type.clone())]);
+        // Replace the update in the wft with a dummy one, since we must instantiate the machines
+        // with the update.
+        let history_update = mem::replace(&mut pwft.work.update, HistoryUpdate::dummy());
+        let mut mrh = ManagedRun::new(
             history_update,
             self.namespace.clone(),
-            workflow_id.to_string(),
-            wf_type.to_string(),
-            run_id.to_string(),
-            metrics.clone(),
-        );
-        let mut mrh = ManagedRunHandle::new(
-            wfm,
+            pwft.work.execution.workflow_id.clone(),
+            pwft.work.workflow_type.clone(),
+            pwft.work.execution.run_id.clone(),
             self.local_activity_request_sink.clone(),
             self.heartbeat_tx.clone(),
             metrics,
         );
-        let rur = mrh.incoming_wft(NewIncomingWFT {
-            history_update: None,
-            start_time,
-        });
-        if self.runs.push(run_id.to_string(), mrh).is_some() {
+        let run_id = run_id.to_string();
+        let rur = mrh.incoming_wft(pwft);
+        if self.runs.push(run_id.clone(), mrh).is_some() {
             panic!("Overflowed run cache! Cache owner is expected to avoid this!");
         }
         self.metrics.cache_size(cur_num_cached_runs as u64 + 1);
-        // This is safe, we just inserted.
-        (self.runs.get_mut(run_id).unwrap(), rur)
+        rur
     }
-    pub fn remove(&mut self, k: &str) -> Option<ManagedRunHandle> {
+    pub fn remove(&mut self, k: &str) -> Option<ManagedRun> {
         let r = self.runs.pop(k);
         self.metrics.cache_size(self.len() as u64);
         r
     }
 
-    pub fn get_mut(&mut self, k: &str) -> Option<&mut ManagedRunHandle> {
+    pub fn get_mut(&mut self, k: &str) -> Option<&mut ManagedRun> {
         self.runs.get_mut(k)
     }
-    pub fn get(&mut self, k: &str) -> Option<&ManagedRunHandle> {
+    pub fn get(&mut self, k: &str) -> Option<&ManagedRun> {
         self.runs.get(k)
     }
 
@@ -125,16 +106,16 @@ impl RunCache {
         self.runs.peek_lru().map(|(run_id, _)| run_id.as_str())
     }
     /// Returns an iterator yielding cached runs in LRU order
-    pub fn runs_lru_order(&self) -> impl Iterator<Item = (&str, &ManagedRunHandle)> {
+    pub fn runs_lru_order(&self) -> impl Iterator<Item = (&str, &ManagedRun)> {
         self.runs.iter().rev().map(|(k, v)| (k.as_str(), v))
     }
-    pub fn peek(&self, k: &str) -> Option<&ManagedRunHandle> {
+    pub fn peek(&self, k: &str) -> Option<&ManagedRun> {
         self.runs.peek(k)
     }
     pub fn has_run(&self, k: &str) -> bool {
         self.runs.contains(k)
     }
-    pub fn handles(&self) -> impl Iterator<Item = &ManagedRunHandle> {
+    pub fn handles(&self) -> impl Iterator<Item = &ManagedRun> {
         self.runs.iter().map(|(_, v)| v)
     }
     pub fn is_full(&self) -> bool {
